make http://validator.w3.org/ happy
Still need some overall fixes for validity, as some of the code is just
"accidentally" valid, and not actually. For example paragraph `<p>` tags
cannot contain block elements (see [1]), such as div (we have that a lot),
just things like span, a, and img tags.

Other things like encding "&" in the URL as "&amp;" is a bit ambiguous,
many people say it's required, others say that it goes agains what
browsers are doing and current validators are out of date. Keep it like
this for the moment.

[1]: http://www.w3.org/TR/html401/struct/text.html#h-9.3.1

Author: Kai Carver <kaicarver@gmail.com>
Amended-by: Gergely Imreh <imrehg@gmail.com>